requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.23.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
-e git+https://github.com/mattaq31/GelGenie.git@23fae2fb6ab2ade7149f4e7a76248dba7d9d23a9#egg=GelGenie&subdirectory=python-gelgenie
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
imagecodecs==2026.1.14
ImageIO==2.37.2
Jinja2==3.1.6
jsonpickle==4.1.1
kiwisolver==1.4.9
lazy_loader==0.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
monai==1.5.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numpy==2.3.5
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==3.0.0
pillow==12.0.0
platformdirs==4.8.0
propcache==0.4.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
rich-click==1.9.7
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.17.0
segmentation_models_pytorch==0.5.0
sentry-sdk==2.52.0
setuptools==82.0.0
shellingham==1.5.4
simsimd==6.5.12
six==1.17.0
smmap==5.0.2
sqlparse==0.5.5
stringzilla==4.6.0
sympy==1.14.0
tifffile==2026.1.28
timm==1.0.24
toml==0.10.2
torch==2.9.1
torchinfo==1.8.0
torchshow==0.5.2
torchvision==0.24.1
tqdm==4.67.3
typer==0.23.1
typer-slim==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wandb==0.25.0
yarl==1.22.0

# --- Core Image / Scientific Stack ---
numpy>=2.0
pillow>=12.0
opencv-python>=4.12

# --- PyTorch Stack (Python 3.12 compatible) ---
torch>=2.6
torchvision>=0.21

# --- Database ---
psycopg2-binary>=2.9.11
psycopg2-pool>=1.2

# --- Runtime Utilities ---
Jinja2>=3.1.6
MarkupSafe>=2.1
filelock>=3.20
fsspec>=2024.6
networkx>=3.5
mpmath>=1.3
sympy>=1.14

sqlparse>=0.5.5
setuptools>=80.9
wheel>=0.45
